fix: 레이싱카 게임뿐만 아니라, 추후 다른 자동차게임도 구현가능하도록 interface상속
